Remove socketserver and use YAML
Radically rewrite server code to remove dependency on socketserver.
Instead, make a SlowControlServer class that executes the main loop and
outsources execution of device commands to separate classes and
communication with the user to a UserHandler class, which uses select to
handle reading and writing from the input and output. Also, replace INI
config with the more powerful YAML format.